Separate corpus report tallying
Amp-Thread-ID: https://ampcode.com/threads/T-01a0609b-6f8d-7718-90c2-86334dba3464
